% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blaster-package.R
\docType{package}
\name{blaster-package}
\alias{blaster}
\alias{blaster-package}
\title{blaster: Native R Implementation of an Efficient BLAST-Like Algorithm}
\description{
Implementation of an efficient BLAST-like sequence comparison algorithm, written in 'C++11' and using native R datatypes. Blaster is based on 'nsearch' - Schmid et al (2018) \doi{10.1101/399782}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tamminenlab/blaster}
}

}
\author{
\strong{Maintainer}: Manu Tamminen \email{mavatam@utu.fi} (\href{https://orcid.org/0000-0001-5891-7653}{ORCID})

Authors:
\itemize{
  \item Timothy Julian \email{tim.julian@eawag.ch} (\href{https://orcid.org/0000-0003-1000-0306}{ORCID})
  \item Aditya Jeevennavar \email{aditya.a.jeevannavar@utu.fi} (\href{https://orcid.org/0000-0002-0737-7316}{ORCID})
  \item Steven Schmid \email{stevschmid@gmail.com}
}

}
\keyword{internal}
