% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pre_blast}
\alias{pre_blast}
\title{Perform a BLAST-like sequence similarity search}
\usage{
pre_blast(
  query_table,
  db_table,
  output_file,
  maxAccepts = 1L,
  maxRejects = 16L,
  minIdentity = 0.75,
  strand = "both"
)
}
\arguments{
\item{query_table}{A DataFrame}

\item{db_table}{A DataFrame}

\item{output_file}{A string}

\item{maxAccepts}{An integer}

\item{maxRejects}{An integer}

\item{minIdentity}{A double}

\item{strand}{A string}
}
\description{
Perform a BLAST-like sequence similarity search
}
