% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{read_dna_fasta}
\alias{read_dna_fasta}
\title{Read the contents of a Fasta file into a DataFrame}
\usage{
read_dna_fasta(filename, filter = "", non_nucleotide_chars = "error")
}
\arguments{
\item{filename}{A string; name of the imported Fasta file.}

\item{filter}{A string; only include those entries which contain a sequence motif specified in this argument.}
}
\value{
A dataframe with Id and Seq columns
}
\description{
Read the contents of a Fasta file into a DataFrame
}
\examples{

query <- system.file("extdata", "query.fasta", package = "blaster")
db <- system.file("extdata", "db.fasta", package = "blaster")

query <- read_dna_fasta(filename = query)

query <- read_dna_fasta(filename = db, filter = "TGGTTGAGG")

}
