% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{read_protein_fasta}
\alias{read_protein_fasta}
\title{Read the contents of a protein Fasta file into a DataFrame}
\usage{
read_protein_fasta(filename, filter = "", non_aa_chars = "error")
}
\arguments{
\item{filename}{A string; name of the imported protein Fasta file.}

\item{filter}{A string; only include those entries which contain a sequence motif specified in this argument.}
}
\value{
A dataframe with Id and Seq columns
}
\description{
Read the contents of a protein Fasta file into a DataFrame
}
\examples{

query <- system.file("extdata", "query.fasta", package = "blaster")
db <- system.file("extdata", "db.fasta", package = "blaster")

}
