% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blaster.R
\name{blast}
\alias{blast}
\title{Runs BLAST algorithm}
\usage{
blast(
  query,
  db,
  maxAccepts = 1,
  maxRejects = 16,
  minIdentity = 0.75,
  alphabet = "nt",
  strand = "both",
  output_to_file = FALSE
)
}
\arguments{
\item{query}{A dataframe or a string}

\item{db}{A dataframe or a string}

\item{maxAccepts}{A number}

\item{maxRejects}{A number}

\item{minIdentity}{A number}

\item{alphabet}{A string}

\item{strand}{A string}

\item{output_to_file}{A boolean}
}
\value{
A dataframe or a string
}
\description{
Runs BLAST algorithm
}
\examples{

query <- system.file("extdata", "query.fasta", package = "blaster")
db <- system.file("extdata", "db.fasta", package = "blaster")

blast_table <- blast(query = query, db = db)

query <- read_fasta(filename = query)
db <- read_fasta(filename = db)
blast_table <- blast(query = query, db = db)

}
