% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blaster.R
\docType{package}
\name{blaster}
\alias{blaster}
\title{Blaster}
\description{
Blaster implements an efficient BLAST-like sequence
comparison algorithm using native R datatypes.
}
\author{
Manu Tamminen <mavatam.@utu.fi>, Steven Schmid
}
