% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{read_fasta}
\alias{read_fasta}
\title{Read the contents of a Fasta file into a DataFrame}
\usage{
read_fasta(filename, filter = "")
}
\arguments{
\item{filename}{A string}

\item{filter}{A string}
}
\description{
Read the contents of a Fasta file into a DataFrame
}
